+++ packages/router/src/Router.ts
@@ -2,6 +2,7 @@ import { Context } from './Context.js';
 import { Dispatcher } from './Dispatcher.js';
 import { Middleware } from './Middleware.js';
 import { Route, RouteFn } from './Route.js';
+import { RouterError } from './RouterError.js';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export class Router<TState extends object = any> implements Dispatcher<TState> {
@@ -30,6 +31,10 @@ export class Router<TState extends object = any> implements Dispatcher<TState> {
   private invokeMiddlewareChain(ctx: Context<TState>, core: Middleware<TState>): Promise<void> {
     const middlewares = this.middlewares;
     const dispatch = (i: number): Promise<void> => {
+      if (ctx.halted) {
+        console.error('race condition or might be invalid next invocation after context halted');
+        return Promise.resolve();
+      }
       const fn = i === middlewares.length ? core : middlewares[i];
       return fn(ctx, () => dispatch(i + 1));
     };