+++ packages/mediator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @xlit/mediator
 
+## 2.6.14
+
+### Patch Changes
+
+- router navigation will halt old context
+
 ## 2.6.13
 
 ### Patch Changes